+++ src/Views/src/Engine/Native/NativeEngine.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Views\Engine\Native;
@@ -20,39 +13,25 @@ final class NativeEngine extends AbstractEngine
 {
     protected const EXTENSION = 'php';
 
-    /** @var ContainerInterface */
-    private $container;
-
-    /**
-     * NativeEngine constructor.
-     */
-    public function __construct(ContainerInterface $container, string $extension = self::EXTENSION)
-    {
-        $this->container = $container;
+    public function __construct(
+        private readonly ContainerInterface $container,
+        string $extension = self::EXTENSION
+    ) {
         $this->extension = $extension;
     }
 
-    /**
-     * @inheritdoc
-     */
     public function compile(string $path, ContextInterface $context): void
     {
         // doing nothing, native views can not be compiled
     }
 
-    /**
-     * @inheritdoc
-     */
     public function reset(string $path, ContextInterface $context): void
     {
         // doing nothing, native views can not be compiled
     }
 
-    /**
-     * @inheritdoc
-     */
     public function get(string $path, ContextInterface $context): ViewInterface
     {
-        return new NativeView($this->getLoader()->load($path), $this->container, $context);
+        return new NativeView($this->getLoader()->load($path), $this->container);
     }
 }