+++ src/Views/src/Processor/ContextProcessor.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Views\Processor;
@@ -23,21 +16,19 @@ final class ContextProcessor implements ProcessorInterface
     // Context injection pattern @{key|default}
     private const PATTERN = '/@\\{(?P<name>[a-z0-9_\\.\\-]+)(?: *\\| *(?P<default>[^}]+))?}/i';
 
-    /** @var string */
-    private $pattern = '';
+    private string $pattern;
 
     public function __construct(string $pattern = null)
     {
         $this->pattern = $pattern ?? static::PATTERN;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function process(ViewSource $source, ContextInterface $context): ViewSource
     {
-        return $source->withCode(preg_replace_callback($this->pattern, static function ($matches) use ($context) {
-            return $context->resolveValue($matches[1]);
-        }, $source->getCode()));
+        return $source->withCode(\preg_replace_callback(
+            $this->pattern,
+            static fn (array $matches) => $context->resolveValue($matches[1]),
+            $source->getCode()
+        ));
     }
 }