+++ src/Views/src/ViewContext.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Views;
@@ -22,19 +15,16 @@ use Spiral\Views\Exception\ContextException;
 final class ViewContext implements ContextInterface
 {
     /** @var DependencyInterface[] */
-    private $dependencies = [];
+    private array $dependencies = [];
 
-    /**
-     * {@inheritdoc}
-     */
     public function getID(): string
     {
         $calculated = '';
         foreach ($this->dependencies as $dependency) {
-            $calculated .= "[{$dependency->getName()}={$dependency->getValue()}]";
+            $calculated .= \sprintf('[%s=%s]', $dependency->getName(), $dependency->getValue());
         }
 
-        return md5($calculated);
+        return \md5($calculated);
     }
 
     /**
@@ -47,9 +37,6 @@ final class ViewContext implements ContextInterface
         return \array_values($this->dependencies);
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function withDependency(DependencyInterface $dependency): ContextInterface
     {
         $environment = clone $this;
@@ -58,13 +45,10 @@ final class ViewContext implements ContextInterface
         return $environment;
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function resolveValue(string $dependency)
+    public function resolveValue(string $dependency): mixed
     {
         if (!isset($this->dependencies[$dependency])) {
-            throw new ContextException("Undefined context dependency '{$dependency}'.");
+            throw new ContextException(\sprintf('Undefined context dependency \'%s\'.', $dependency));
         }
 
         return $this->dependencies[$dependency]->getValue();