+++ src/Views/src/ViewInterface.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Views;
@@ -17,7 +10,6 @@ interface ViewInterface
 {
     /**
      * Render view source using internal logic.
-     *
      *
      * @throws RenderException
      */