+++ src/Views/src/ViewManager.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Views;
@@ -17,24 +10,18 @@ use Spiral\Views\Exception\ViewException;
 
 final class ViewManager implements ViewsInterface
 {
-    /** @var ViewsConfig */
-    private $config;
-
-    /** @var ContextInterface */
-    private $context;
-
-    /** @var LoaderInterface */
-    private $loader;
-
-    /** @var ViewCache|null */
-    private $cache;
+    private readonly LoaderInterface $loader;
+    private ?ViewCache $cache = null;
+    private ContextInterface $context;
 
     /** @var EngineInterface[] */
-    private $engines = [];
+    private array $engines = [];
 
-    public function __construct(ViewsConfig $config, FactoryInterface $factory, ?ContextInterface $context = null)
-    {
-        $this->config = $config;
+    public function __construct(
+        private readonly ViewsConfig $config,
+        FactoryInterface $factory,
+        ?ContextInterface $context = null
+    ) {
         $this->context = $context ?? new ViewContext();
         $this->loader = $factory->make(LoaderInterface::class, [
             'namespaces' => $config->getNamespaces(),
@@ -98,9 +85,7 @@ final class ViewManager implements ViewsInterface
      */
     public function compile(string $path): void
     {
-        if ($this->cache !== null) {
-            $this->cache->resetPath($path);
-        }
+        $this->cache?->resetPath($path);
 
         $engine = $this->findEngine($path);
 
@@ -117,9 +102,7 @@ final class ViewManager implements ViewsInterface
      */
     public function reset(string $path): void
     {
-        if ($this->cache !== null) {
-            $this->cache->resetPath($path);
-        }
+        $this->cache?->resetPath($path);
 
         $engine = $this->findEngine($path);
 
@@ -133,26 +116,22 @@ final class ViewManager implements ViewsInterface
     /**
      * Get view from one of the associated engines.
      *
-     *
      * @throws ViewException
      */
     public function get(string $path): ViewInterface
     {
-        if ($this->cache !== null && $this->cache->has($this->context, $path)) {
+        if ($this->cache?->has($this->context, $path)) {
             return $this->cache->get($this->context, $path);
         }
 
         $view = $this->findEngine($path)->get($path, $this->context);
 
-        if ($this->cache !== null) {
-            $this->cache->set($this->context, $path, $view);
-        }
+        $this->cache?->set($this->context, $path, $view);
 
         return $view;
     }
 
     /**
-     *
      * @throws ViewException
      */
     public function render(string $path, array $data = []): string
@@ -172,6 +151,6 @@ final class ViewManager implements ViewsInterface
             }
         }
 
-        throw new ViewException("Unable to detect view engine for `{$path}`.");
+        throw new ViewException(\sprintf('Unable to detect view engine for `%s`.', $path));
     }
 }